pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "task-pipeliner"
version = "0.2.0"
description = "Python pipeline execution framework for parallel and sequential data processing"
requires-python = ">=3.12"
dependencies = [
    "PyYAML>=6.0.2",
    "pydantic>=2.0",
    "click>=8.1",
]

[project.scripts]
task-pipeliner = "task_pipeliner.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-timeout>=2.3",
    "pytest-cov>=5.0",
    "mypy>=1.8",
    "types-PyYAML>=6.0",
    "ruff>=0.4",
    "orjson>=3.9",
]

[tool.hatch.build.targets.wheel]
packages = ["src/task_pipeliner"]

[tool.pytest.ini_options]
testpaths = ["tests"]
timeout = 30
pythonpath = ["src", "tests"]

[tool.mypy]
python_version = "3.12"
strict = true
files = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
